refactor: remove audit log export common package
