+++ pkg/storage/apis/obot.obot.ai/v1/auditlogexport.go
@@ -22,22 +22,6 @@ type AuditLogExport struct {
 	Status AuditLogExportStatus `json:"status,omitempty"`
 }
 
-func (a *AuditLogExport) Bucket() string {
-	return a.Spec.Bucket
-}
-
-func (a *AuditLogExport) KeyPrefix() string {
-	return a.Spec.KeyPrefix
-}
-
-func (a *AuditLogExport) SpecName() string {
-	return a.Spec.Name
-}
-
-func (a *AuditLogExport) ExportStatus() *AuditLogExportStatus {
-	return &a.Status
-}
-
 func (a *AuditLogExport) Has(field string) (exists bool) {
 	return slices.Contains(a.FieldNames(), field)
 }
